requirements.txt
fastapi>=0.115.0
playwright>=1.52.0
pydantic>=2.9.0
python-dotenv>=1.0.1
uvicorn[standard]>=0.30.0
